Move mruby dependent codes

+++ include/co5300.h
@@ -1,19 +1,21 @@
 #ifndef CO5300_DEFINED_H_
 #define CO5300_DEFINED_H_
 
+#include <stdint.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
 
 typedef struct co5300_config {
-  uint pwr_pin;
-  uint rst_pin;
-  uint cs_pin;
-  uint sclk_pin;
-  uint d0_pin;
-  uint pio_num;
-  uint sm;
-  uint offset;
+  uint8_t pwr_pin;
+  uint8_t rst_pin;
+  uint8_t cs_pin;
+  uint8_t sclk_pin;
+  uint8_t d0_pin;
+  uint8_t pio_num;
+  uint8_t sm;
+  uint8_t offset;
 } co5300_config_t;
 
 void CO5300_init(co5300_config_t *config);
